backend/requirements.txt
alembic==1.14.1
email-validator==2.2.0
fastapi==0.115.6
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
PyJWT==2.9.0
python-multipart==0.0.20
SQLAlchemy==2.0.36
uvicorn[standard]==0.33.0
